+++ LOC/FR/strings_db.lua
@@ -7095,6 +7095,8 @@ tooltipui0686="Pas de Billy"
 tooltipui0687="Empêche les commandants UEF d'améliorer leur ACU avec un missile nucléaire tactique 'Billy'"
 tooltipui0688="Pas de super structures d'espionage"
 tooltipui0689="Pas d'Oeil de Rihanne (Aeon) ni de Devin (Cybran)"
+tooltipui0690="Drone Auto-Rebuild"
+tooltipui0691="Toggle if this drone rebuilds upon death"
 
 tooltipui0540="Réinitialiser la fenêtre de discussion"
 tooltipui0541="Réinitialise la position et la disposition de la fenêtre de discussion."